refactor(ivy): move styling files around (#26096)
PR Close #26096
